requirements.txt
boto3
urllib3
botocore
scipy
shellingham
simple-websocket
six
sniffio
sortedcontainers
SQLAlchemy
starlette
tabulate
tenacity
threadpoolctl
tiktoken
together
tqdm
trio
typer
typing-inspect
typing_extensions
tzdata
ujson
uvicorn
uvloop
watchfiles
websocket-client
websockets
Werkzeug
wsproto
xgboost
yarl
zope.event
zope.interface
aiofiles
aiohttp
aiosignal
annotated-types
anyio
async-timeout
attrs
bidict
blinker
certifi
charset-normalizer
click
contourpy
cycler
dataclasses-json
distro
dnspython
email_validator
eval_type_backport
exceptiongroup
fastapi
fastapi-cli
filelock
Flask
Flask-SocketIO
fonttools
frozenlist
gevent
greenlet
h11
httpcore
httptools
httpx
idna
itsdangerous
Jinja2
jmespath
joblib
jsonpatch
jsonpointer
kiwisolver
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langchain-together
langsmith
markdown-it-py
MarkupSafe
marshmallow
matplotlib
mdurl
multidict
mypy-extensions
numpy
openai
orjson
outcome
packaging
pandas
pillow
pinecone
pyarrow
pydantic
pydantic_core
Pygments
pyparsing
python-dateutil
python-dotenv
python-engineio
python-multipart
python-socketio
pytz
PyYAML
regex
requests
rich
s3transfer
scikit-learn
fastapi
socketio
python-socketio[asgi]
uvicorn
boto3
pymongo
jupyter_client
notebook
jupyterlab
jupyter
beautifulsoup4
pymongo
motor